+++ src/index.ts
@@ -1,7 +1,8 @@
-import * as fs from 'fs';
+
 
 import { initGUI } from './guiLoader'
 import { generateQR } from './qrCodeGenerator'
+import { configureExpress } from './webserver'
 
 import sqlite3 from 'sqlite3'
 import { open } from 'sqlite'
@@ -25,34 +26,8 @@ import {MediaInfo} from './mediainfo'
 
   // rest of the code remains same
   const app = express();
-
-  app.use(express.static('public'));
-  app.get('/', (req, res) => {
-    const files: any[] = [];
-    fs.readdirSync('public/media').forEach(file => {
-      if (file.endsWith(".webm")) {
-        const info:MediaInfo = JSON.parse(fs.readFileSync("public/media/" + file.replace(".webm", ".json")).toString())
-
-        files.push({ filename: file, title: info.title });
-      }
-
-    });
-    const fileLinks: string = files.map(x => `<a href="video.html?v=${x.filename}">${x.title}</a>`).join("<br/>")
-    res.send(fileLinks)
-  });
-  app.get('/mediainfo', (req, res) => {
-
-    const infos: any[] = [];
-    fs.readdirSync('public/media').forEach(file => {
-      if (file.endsWith(".webm")) {
-        const info:MediaInfo = JSON.parse(fs.readFileSync("public/media/" + file.replace(".webm", ".json")).toString())
-
-        infos.push(info);
-      }
-
-    });res.send(infos);
-
-  });
+configureExpress(app);
+  
   app.listen(port, () => {
     console.log(`⚡️[server]: Server is running at https://localhost:${port}`);
   });
